+++ src/test/java/com/ttasjwi/cafe/domain/article/MemoryArticleRepositoryTest.java
@@ -21,7 +21,11 @@ class MemoryArticleRepositoryTest {
     @DisplayName("저장 게시글의 인덱스로 조회 -> 같은 저장게시글 반환")
     void saveTest() {
         // given
-        Article article = new Article("자바 두명 타요", "활활 잘 탄다~");
+        Article article = Article.builder()
+                .title("자바 두명 타요")
+                .content("활활 잘 탄다~")
+                .build();
+
         memoryArticleRepository.save(article);
         Long articleId = article.getArticleId();
 
@@ -36,8 +40,15 @@ class MemoryArticleRepositoryTest {
     @DisplayName("findAll -> 저장한 인원만큼의 리스트 반환")
     void findAllTest() {
         // given
-        Article article1 = new Article("선생님들 살려주세요", "회사 DB에 DROP DATABASE 해버렸어요;;");
-        Article article2 = new Article("오늘의 점심 추천 받습니다...", "ㅜㅜ");
+        Article article1 = Article.builder()
+                .title("선생님들 살려주세요")
+                .content("회사 DB에 DROP DATABASE 해버렸어요;;")
+                .build();
+
+        Article article2 = Article.builder()
+                .title("우리 개발자 연봉 이대로 가면")
+                .content("라이더")
+                .build();
 
         memoryArticleRepository.save(article1);
         memoryArticleRepository.save(article2);